create renderImage function and move it to separate file
